+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `MUNICIPALITY` added to `Precision` enum
 - `requests-cache` added as optional dependency
+- `excel_str_german` property added to `Detail`
 
 ### Changed